refactor: finalize useCriteriaManager composable extraction
